chore: fix clippy warnings, doctests, and update README
- Fix all clippy -D warnings across entire workspace:
  - Remove assert!(true) patterns on constant expressions
  - Fix uninlined format args in format strings
  - Replace unnecessary casts and useless conversions
  - Add #[allow(dead_code)] for feature-gated modules
  - Fix boolean assertion patterns and unused variables
  - Use div_ceil() instead of manual division
  - Use Error::other() instead of Error::new()
  - Fix single_match clippy lint in sendfile
  - Remove useless vec! and assert! patterns
- Fix broken doctests across cli, engine, protocol crates:
  - Update dry_run imports to use public re-exports
  - Update sparse/partial imports to use crate-level re-exports
  - Fix Vec type inference in MplexReader examples
  - Mark platform-dependent examples as no_run
- Fix rustdoc broken intra-doc link in daemon auth module
- Remove "none" from --version algorithm lists
- Update README with accurate test count (21,649+) and crate list
- Full validation passes: fmt, clippy -D warnings, nextest, xtask docs

+++ crates/flist/src/batched_stat.rs
@@ -407,19 +407,19 @@ impl StatxResult {
     /// Returns true if this entry is a regular file.
     #[must_use]
     pub fn is_file(&self) -> bool {
-        (self.mode & libc::S_IFMT as u32) == libc::S_IFREG as u32
+        (self.mode & libc::S_IFMT) == libc::S_IFREG
     }
 
     /// Returns true if this entry is a directory.
     #[must_use]
     pub fn is_dir(&self) -> bool {
-        (self.mode & libc::S_IFMT as u32) == libc::S_IFDIR as u32
+        (self.mode & libc::S_IFMT) == libc::S_IFDIR
     }
 
     /// Returns true if this entry is a symbolic link.
     #[must_use]
     pub fn is_symlink(&self) -> bool {
-        (self.mode & libc::S_IFMT as u32) == libc::S_IFLNK as u32
+        (self.mode & libc::S_IFMT) == libc::S_IFLNK
     }
 
     /// Returns the permission bits (lower 12 bits of mode).
